Update LineNumber.java
Created a List<String> lines to store each line (for Proof.print()). Everytime ln.add() or ln.addPoint() is called, lines.add(toString()) adds the new LineNumber string to the ArrayList of Lines.

Wrote isLegal to check for any non integer or non ".".

Overloaded isLegal to take in a string and the current LineNumber. Created a function lineLevel to count the number of "." in a LineNumber. If the input string to isLegal has more "." than the current LineNumber, return false.