Update Proof.java
So, apparently we are only allowed to call "assume" directly after a "show". In some testing, we call multiple "assumes" so that we can test our functions. So I added a isDebugging boolean.

Added else if case to extendProof() when user calls a theorem name. Checks that the expression is legal, then compares the expression to the theorem's expression. If they are compatible, set the expression to be a fact.